data-ingester/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
requests
httpx
python-dotenv
